Removed all logic from TagManager and added HA search
Searching keyword changed to HA: and GE: respectively
